Move fixtures into their own file
